can {fpc}	R Documentation

Generation of the tuning constant for regression fixed point clusters

Description

Generates tuning constants ca for fixreg dependent on the number of points and variables of the dataset.

Only thought for use in fixreg.

Usage

can(n, p)

Arguments

n
positive integer. Number of points.

p
positive integer. Number of independent variables.

Details

The formula is 3+33/(n*2^{-(p-1)/2})^{1/3}+2900000/(n*2^{-(p-1)/2})^3. For justification cf. Hennig (2002).

Value

A number.

Author(s)

Christian Hennig c.hennig@ucl.ac.uk http://www.homepages.ucl.ac.uk/~ucakche/

References

Hennig, C. (2002) Fixed point clusters for linear regression: computation and comparison, Journal of Classification 19, 249-276.

See Also

fixreg

Examples

  can(429,3)

[Package fpc version 2.1-11.1 Index]