clusexpect {fpc}	R Documentation

Expected value of the number of times a fixed point cluster is found

Description

A rough approximation of the expectation of the number of times a well separated fixed point cluster (FPC) of size n is found in ir fixed point iterations of fixreg.

Usage

  clusexpect(n, p, cn, ir)

Arguments

n
positive integer. Total number of points.

p
positive integer. Number of independent variables.

cn
positive integer smaller or equal to n. Size of the FPC.

ir
positive integer. Number of fixed point iterations.

Details

The approximation is based on the assumption that a well separated FPC is found iff all p+2 points of the initial coinfiguration come from the FPC. The value is ir times the probability for this. For a discussion of this assumption cf. Hennig (2002).

Value

A number.

Author(s)

Christian Hennig c.hennig@ucl.ac.uk http://www.homepages.ucl.ac.uk/~ucakche/

References

Hennig, C. (2002) Fixed point clusters for linear regression: computation and comparison, Journal of Classification 19, 249-276.

See Also

fixreg

Examples

  round(clusexpect(500,4,150,2000),digits=2)

[Package fpc version 2.1-11.1 Index]