cmahal {fpc}	R Documentation

Generation of tuning constant for Mahalanobis fixed point clusters.

Description

Generates tuning constants ca for fixmahal dependent on the number of points and variables of the current fixed point cluster (FPC).

This is experimental and only thought for use in fixmahal.

Usage

cmahal(n, p, nmin, cmin, nc1, c1 = cmin, q = 1)

Arguments

n
positive integer. Number of points.

p
positive integer. Number of variables.

nmin
integer larger than 1. Smallest number of points for which ca is computed. For smaller FPC sizes, ca is set to the value for nmin.

cmin
positive number. Minimum value for ca.

nc1
positive integer. Number of points at which ca=c1.

c1
positive numeric. Tuning constant for cmahal. Value for ca for FPC size equal to nc1.

q
numeric between 0 and 1. 1 for steepest possible descent of ca as function of the FPC size. Should presumably always be 1.

Details

Some experiments suggest that the tuning constant ca should decrease with increasing FPC size and increase with increasing p in fixmahal. This is to prevent too small meaningless FPCs while maintaining the significant larger ones. cmahal with q=1 computes ca in such a way that as long as ca>cmin, the decrease in n is as steep as possible in order to maintain the validity of the convergence theorem in Hennig and Christlieb (2002).

Value

A numeric vector of length n, giving the values for ca for all FPC sizes smaller or equal to n.

Author(s)

Christian Hennig c.hennig@ucl.ac.uk http://www.homepages.ucl.ac.uk/~ucakche/

References

Hennig, C. and Christlieb, N. (2002) Validating visual clusters in large datasets: Fixed point clusters of spectral features, Computational Statistics and Data Analysis 40, 723-739.

See Also

fixmahal

Examples

  plot(1:100,cmahal(100,3,nmin=5,cmin=qchisq(0.99,3),nc1=90),
       xlab="FPC size", ylab="cmahal")

[Package fpc version 2.1-11.1 Index]