dipp.tantrum {fpc}	R Documentation

Simulates p-value for dip test

Description

Simulates p-value for dip test (see dip) in the way suggested by Tantrum, Murua and Stuetzle (2003) from the clostest unimodal distribution determined by kernel density estimation with bandwith chosen so that the density just becomes unimodal. This is less conservative (and in fact sometimes anti-conservative) than the values from dip.test.

Usage

  dipp.tantrum(xdata,d,M=100)

Arguments

xdata
numeric vector. One-dimensional dataset.

d
numeric. Value of dip statistic.

M
integer. Number of artificial datasets generated in order to estimate the p-value.

Value

List with components

p.value
approximated p-value.

bw
borderline unimodality bandwith in density with default settings.

dv
vector of dip statistic values from simulated artificial data.

Author(s)

Christian Hennig c.hennig@ucl.ac.uk http://www.homepages.ucl.ac.uk/~ucakche/

References

J. A. Hartigan and P. M. Hartigan (1985) The Dip Test of Unimodality, Annals of Statistics, 13, 70-84.

Tantrum, J., Murua, A. and Stuetzle, W. (2003) Assessment and Pruning of Hierarchical Model Based Clustering, Proceedings of the ninth ACM SIGKDD international conference on Knowledge discovery and data mining, Washington, D.C., 197-205.

Examples

# not run, requires package diptest
#  x <- runif(100)
#  d <- dip(x)
#  dt <- dipp.tantrum(x,d,M=10)

[Package fpc version 2.1-11.1 Index]