dudahart2 {fpc}	R Documentation

Duda-Hart test for splitting

Description

Duda-Hart test for whether a data set should be split into two clusters.

Usage

  dudahart2(x,clustering,alpha=0.001)

Arguments

x
data matrix or data frame.

clustering
vector of integers. Clustering into two clusters.

alpha
numeric between 0 and 1. Significance level (recommended to be small if this is used for estimating the number of clusters).

Value

A list with components

p.value
p-value against null hypothesis of homogemeity.

dh
ratio of within-cluster sum of squares for two clusters and overall sum of squares.

compare
critical value for dh at level alpha.

cluster1
FALSE if the null hypothesis of homogemeity is rejected.

alpha
see above.

z
1-alpha-quantile of a standard Gaussian.

Author(s)

Christian Hennig c.hennig@ucl.ac.uk http://www.homepages.ucl.ac.uk/~ucakche

References

Duda, R. O. and Hart, P. E. (1973) Pattern Classification and Scene Analysis. Wiley, New York.

See Also

cluster.stats

Examples

  options(digits=2)
  set.seed(98765)
  iriss <- iris[sample(150,20),-5]
  km <- kmeans(iriss,2)
  dudahart2(iriss,km$cluster)

[Package fpc version 2.1-11.1 Index]