extract.mixturepars {fpc}	R Documentation

Extract parameters for certain components from mclust

Description

Extracts parameter of certain mixture components from the output of summary.mclustBIC and updates proportions so that they sum up to 1.

Usage

  extract.mixturepars(mclustsum,compnumbers,noise=FALSE)

Arguments

mclustsum
output object of summary.mclustBIC.

compnumbers
vector of integers. Numbers of mixture components.

noise
logical. Should be TRUE if a noise component was fitted by mclustBIC.

Value

Object as component parameters of summary.mclustBIC-output, but for specified components only. (Orientation information from all components is kept.)

Author(s)

Christian Hennig c.hennig@ucl.ac.uk http://www.homepages.ucl.ac.uk/~ucakche/

Examples

  set.seed(98765)
  options(digits=2)
  require(mclust)
  iriss <- iris[sample(150,20),-5]
  irisBIC <- mclustBIC(iriss)
  siris <- summary(irisBIC,iriss)
  extract.mixturepars(siris,2)

[Package fpc version 2.1-11.1 Index]