fpclusters {fpc}	R Documentation

Extracting clusters from fixed point cluster objects

Description

fpclusters is a generic function which extracts the representative fixed point clusters (FPCs) from FPC objects generated by fixmahal and fixreg. For documentation and examples see fixmahal and fixreg.

Usage

fpclusters(object, ...)

Arguments

object
object of class rfpc or mfpc.

...
further arguments depending on the method.

Value

a list of logical or numerical vectors indicating or giving the weights of the cluster memberships.

Author(s)

Christian Hennig c.hennig@ucl.ac.uk http://www.homepages.ucl.ac.uk/~ucakche/

See Also

fixmahal, fixreg

[Package fpc version 2.1-11.1 Index]