itnumber {fpc}	R Documentation

Number of regression fixed point cluster iterations

Description

Computes the number of fixed point iterations needed by fixreg to find mtf times a fixed point cluster (FPC) of size cn with an approximated probability of prob.

Thought for use within fixreg.

Usage

itnumber(n, p, cn, mtf, prob = 0.95, maxir = 20000)

Arguments

n
positive integer. Total number of points.

p
positive integer. Number of independent variables.

cn
positive integer smaller or equal to n. Size of the FPC.

mtf
positive integer.

prob
number between 0 and 1.

maxir
positive integer. itnumber is set to this value if it would otherwise be larger.

Details

The computation is based on the binomial distribution with probability given by clusexpect with ir=1.

Value

An integer.

Author(s)

Christian Hennig c.hennig@ucl.ac.uk http://www.homepages.ucl.ac.uk/~ucakche/

References

Hennig, C. (2002) Fixed point clusters for linear regression: computation and comparison, Journal of Classification 19, 249-276.

See Also

fixreg, clusexpect

Examples

  itnumber(500,4,150,2)

[Package fpc version 2.1-11.1 Index]