mahalanodisc {fpc}	R Documentation

Mahalanobis for AWC

Description

Vector of Mahalanobis distances or their root. For use in awcoord only.

Usage

mahalanodisc(x2, mg, covg, modus="square") 

Arguments

x2
numerical data matrix.

mg
mean vector.

covg
covariance matrix.

modus
"md" (roots of Mahalanobis distances) or "square" (original squared form of Mahalanobis distances).

Details

The covariance matrix is inverted by use of solvecov, which can be expected to give reasonable results for singular within-class covariance matrices.

Value

vector of (rooted) Mahalanobis distances.

Author(s)

Christian Hennig c.hennig@ucl.ac.uk http://www.homepages.ucl.ac.uk/~ucakche/

See Also

awcoord, solvecov

Examples

  options(digits=3)
  x <- cbind(rnorm(50),rnorm(50))
  mahalanodisc(x,c(0,0),cov(x))
  mahalanodisc(x,c(0,0),matrix(0,ncol=2,nrow=2))

[Package fpc version 2.1-11.1 Index]