mahalconf {fpc}	R Documentation

Mahalanobis fixed point clusters initial configuration

Description

Generates an initial configuration of startn points from dataset x for the fixmahal fixed point iteration.

Thought only for use within fixmahal.

Usage

mahalconf(x, no, startn, covall, plot)

Arguments

x
numerical matrix. Rows are points, columns are variables.

no
integer between 1 and nrow(x). Number of the first point of the configuration.

startn
integer between 1 and nrow(x).

covall
covariance matrix for the computation of the first Mahalanobis distances.

plot
a string. If equal to "start" or "both",the first two variables and the first ncol(x)+1 points are plotted.

Details

mahalconf first chooses the p (number of variables) nearest points to point no. no in terms of the Mahalanobis distance w.r.t. covall, so that there are p+1 points. In every further step, the covariance matrix of the current configuration is computed and the nearest point in terms of the new Mahalanobis distance is added. solvecov is used to invert singular covariance matrices.

Value

A logical vector of length nrow(x).

Author(s)

Christian Hennig c.hennig@ucl.ac.uk http://www.homepages.ucl.ac.uk/~ucakche/

See Also

fixmahal, solvecov

Examples

  set.seed(4634)
  face <- rFace(600,dMoNo=2,dNoEy=0,p=2)
  mahalconf(face,no=200,startn=20,covall=cov(face),plot="start")

[Package fpc version 2.1-11.1 Index]