minsize {fpc}	R Documentation

Minimum size of regression fixed point cluster

Description

Computes the minimum size of a fixed point cluster (FPC) which is found at least mtf times with approximated probability prob by ir fixed point iterations of fixreg.

Thought for use within fixreg.

Usage

minsize(n, p, ir, mtf, prob = 0.5)

Arguments

n
positive integer. Total number of points.

p
positive integer. Number of independent variables.

ir
positive integer. Number of fixed point iterations.

mtf
positive integer.

prob
numerical between 0 and 1.

Details

The computation is based on the binomial distribution with probability given by clusexpect with ir=1.

Value

An integer.

Author(s)

Christian Hennig c.hennig@ucl.ac.uk http://www.homepages.ucl.ac.uk/~ucakche/

References

Hennig, C. (2002) Fixed point clusters for linear regression: computation and comparison, Journal of Classification 19, 249-276.

See Also

fixreg, clusexpect, itnumber

Examples

  minsize(500,4,7000,2)

[Package fpc version 2.1-11.1 Index]