piridge.zeroes {fpc}	R Documentation

Extrema of two-component Gaussian mixture

Description

By use of the Pi-function in Ray and Lindsay, 2005, locations of two-component Gaussian mixture density extrema or saddlepoints are computed.

Usage

piridge.zeroes(prop, mu1, mu2, Sigma1, Sigma2, alphamin=0,
                          alphamax=1,by=0.001)

Arguments

prop
proportion of mixture component 1.

mu1
mean vector of component 1.

mu2
mean vector of component 2.

Sigma1
covariance matrix of component 1.

Sigma2
covariance matrix of component 2.

alphamin
minimum alpha value.

alphamax
maximum alpha value.

by
interval between alpha-values where to look for extrema.

Value

list with components

number.zeroes
number of zeroes of Pi-function, i.e., extrema or saddlepoints of density.

estimated.roots
estimated alpha-values at which extrema or saddlepoints occur.

Author(s)

Christian Hennig c.hennig@ucl.ac.uk http://www.homepages.ucl.ac.uk/~ucakche/

References

Ray, S. and Lindsay, B. G. (2005) The Topography of Multivariate Normal Mixtures, Annals of Statistics, 33, 2042-2065.

Examples

  q <- piridge.zeroes(0.2,c(1,1),c(2,5),diag(2),diag(2),by=0.1)

[Package fpc version 2.1-11.1 Index]