randcmatrix {fpc}	R Documentation

Random partition matrix

Description

For use within regmix. Generates a random 0-1-matrix with n rows and cln columns so that every row contains exactly one one and every columns contains at least p+3 ones.

Usage

randcmatrix(n,cln,p)

Arguments

n
positive integer. Number of rows.

cln
positive integer. Number of columns.

p
positive integer. See above.

Value

An n*cln-matrix.

Author(s)

Christian Hennig c.hennig@ucl.ac.uk http://www.homepages.ucl.ac.uk/~ucakche/

See Also

regmix

Examples

set.seed(111)
randcmatrix(10,2,1)

[Package fpc version 2.1-11.1 Index]