ridgeline {fpc}	R Documentation

Ridgeline computation

Description

Computes (alpha*Sigma1^-1+(1-alpha)*Sigma2^-1)^-1* alpha*(Sigma1^-1*mu1)+(1-alpha)*(Sigma2^-1*mu2)) as required for the computation of the ridgeline (Ray and Lindsay, 2005) to find all density extrema of a two-component Gaussian mixture with mean vectors mu1 and mu2 and covariance matrices Sigma1, Sigma2.

Usage

ridgeline(alpha, mu1, mu2, Sigma1, Sigma2)

Arguments

alpha
numeric between 0 and 1.

mu1
mean vector of component 1.

mu2
mean vector of component 2.

Sigma1
covariance matrix of component 1.

Sigma2
covariance matrix of component 2.

Value

A vector. See above.

Author(s)

Christian Hennig c.hennig@ucl.ac.uk http://www.homepages.ucl.ac.uk/~ucakche/

References

Ray, S. and Lindsay, B. G. (2005) The Topography of Multivariate Normal Mixtures, Annals of Statistics, 33, 2042-2065.

Examples

  ridgeline(0.5,c(1,1),c(2,5),diag(2),diag(2))

[Package fpc version 2.1-11.1 Index]