sseg {fpc}	R Documentation

Position in a similarity vector

Description

sseg(i,j) gives the position of the similarity of objects i and j in the similarity vectors produced by fixreg and fixmahal. sseg should only be used as an auxiliary function in fixreg and fixmahal.

Usage

sseg(i, j)

Arguments

i
positive integer.

j
positive integer.

Value

A positive integer.

Author(s)

Christian Hennig c.hennig@ucl.ac.uk http://www.homepages.ucl.ac.uk/~ucakche/

Examples

sseg(3,4)

[Package fpc version 2.1-11.1 Index]