tdecomp {fpc}	R Documentation

Root of singularity-corrected eigenvalue decomposition

Description

Computes transposed eigenvectors of matrix m times diagonal of square root of eigenvalues so that eigenvalues smaller than 1e-6 are set to 1e-6.

Usage

  tdecomp(m)

Arguments

m
a symmetric matrix of minimum format 2*2.

Details

Thought for use in discrcoord only.

Value

a matrix.

Note

Thought for use within discrcoord only.

Author(s)

Christian Hennig c.hennig@ucl.ac.uk http://www.homepages.ucl.ac.uk/~ucakche/

Examples

x <- rnorm(10)
y <- rnorm(10)
z <- cov(cbind(x,y))
round(tdecomp(z),digits=2)

[Package fpc version 2.1-11.1 Index]