unimodal.ind {fpc}	R Documentation

Is a fitted denisity unimodal or not?

Description

Checks whether a series of fitted density values (such as given out as y-component of density) is unimodal.

Usage

  unimodal.ind(y)

Arguments

y
numeric vector of fitted density values in order of increasing x-values such as given out as y-component of density.

Value

Logical. TRUE if unimodal.

Author(s)

Christian Hennig c.hennig@ucl.ac.uk http://www.homepages.ucl.ac.uk/~ucakche/

Examples

unimodal.ind(c(1,3,3,4,2,1,0,0))

[Package fpc version 2.1-11.1 Index]